flask_backend/requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-decouple==3.8
redis==6.2.0
requests==2.32.4
spotipy==2.25.1
urllib3==2.5.0
Werkzeug==3.1.3
